add getChildren method for VictoryStack
